Implements support for handling multiple collision types.
The interface for collision systems is changed.
Instead of using "collisionsystem_sprite_data_create"
for creating a collision mask,
the function "get_collision_mask" is used to
fetch a collision mask, and "free_collision_mask"
to free the mask.
The interfaces for the 3 current collision systems,
None, BBox and Precise has been updated and verified to compile
and work.

The interface is also augmented with an enum for collision types
in "Collision_Systems/collision_types.h".
The order of the enums fits with the MaskShape enum in
"org.lateralgm.resources.Sprite" for easy conversion.

The compiler has been augmented to write in a number for the
given collision type in the game, and "Universal_System/spriteinit.cpp"
has been augmented to read and handle the collision type.

The interface for "sprite_set_subimage" in "Universal_System/spritestruct.{cpp,h}"
has been changed to support different collision types,
and functions affected by the interface change has been updated.

The Precise collision system has been augmented with partial support
for different collisions, currently supporting bounding boxes and precise,
and it has been verified to work.