Fixes lighting for positional lights in d3d and fixes normals.
The positional lights were broken before, and were treated
as directional lights due to a missing fourth argument.
Their position was not updated properly,
which meant that a light once defined would stay at that
positive relative to the viewer, not the world.
The position is now stored for positional lights,
and the position is updated at the start of redraw_screen.

The range of positional lights is handled by using attenuation.
The attenuation is based off of the range argument given
when creating positional lights.

The normals for walls are computed correctly,
and the normals for floors have been improved,
though are not correct for slanted floors.