Fixed repeated drawing bug when changing depth during draw event.
The fix is done by delaying any changes to depth through assignment
in a map from instance ids to depth changes.
The map is iterated through before draw events are handled,
and then cleared.
Any changes to the depth during any event, including the draw event, will be
stored in the depth variable and as a change in the map, and not directly
stored in the depth structure.