Implemented automatic collision updating for Box2D.
The interface also allows other collision systems
to use automatic updating. The updating happens before
the collision events are executed, if any.

The changes have been tested and verified to work
for all current collision systems.
For Box2D, remember to remove any calls to
"physics_world_update" that was used to handle the
absence of automatic collision updating.