Moved user constants, variables and functions to enigma_user namespace.
Most of the user parts in Extensions was moved,
and some user parts in Universal was moved as well.
All of the moved parts have been tested and verified
to compile, and some of the functions have been tested
to work correctly, including alarms, StudioPhysics,
DateTime, MotionPlanning, DataStructures and others.

Notable exception for moving is ParticleSystems, which
is going to be looked at later.