Moved things to enigma_user; cleaned up mathnc.{cpp,h}.
Various constants, variables and functions in Universal_System
has been moved to namespace enigma_user. The changes have been
generally tested and verified to work.

mathnc.{cpp,h} have been refactored, cleaned up, and various
things have been moved to enigma_user. The move to enigma_user
is not fully done, since they need to be done simultaneously
with the user space move to enigma_user.
The overloaded math functions for 'var' and 'variant' have been
mostly removed, and it has been tested for both floating point
literals and variables that the changes work. The overloaded
function for 'var' and 'variant' that has not been removed yet
is 'abs'. The bessel functions may not work currently. The
y{0,1,n} and j{0,1,n} names has been tested and verified to be
usable for variables in games.