Moves the particle graphics bridges to Particles.
The graphics bridges are contained in header files,
such that they are only compiled when the
corresponding graphics systems are used.
Currently, OpenGL1 and OpenGL3 have bridges
provided. This ensures that a graphics bridge
that is incompatible with a given graphics system
(like graphics bridge OpenGL3 with graphics system
OpenGL1) are not compiled together.

The graphics bridges are written purely in headers.
This is done since .cpp source files are compiled
before knowledge which graphics system has been
chosen is available.

The OpenGL3 graphics bridge has furthermore been
updated to use OpenGL 3.0+ features, such as shaders
and VBOs.