Added fallback bridge to particles.
The fallback bridge ensures that the particles implementation
works for any graphics system that supports the standard
drawing functions, even if there doesn't exist a specialized
bridge for that graphics system yet.
The bridge will generally not be as fast as it could be.

The fallback bridge does currently not support blending, due to the
standard drawing functions not being sufficient for implementing
it properly. This is considered OK, since partly the fallback
bridge is a fallback system, and partly because it is very easy
to specialize the fallback bridge and implement blending correctly.
The fallback bridge should work correctly in all other regards.

Also made minor refactoring and improvements to texture handling
in particles.