Refactored and improved compiling games from the command line.
Changed the return status code from "double" to "int" in the
compiler, and made the plugin exit with that code when compiling.

Improved the interface and functionality of the plugin when
compiling. A help interface is printed with -h, and the plugin
now supports an argument to let the user select the API,
overriding the game's selection. This is useful in cases such
as when doing automated integration testing for several different
API selections. An argument for the output location of the
compiled executable is also supported.

Also fixed a minor bug in EnigmaSettings.java.