Added integration testing library, tests, format, information.
Also added missing header to the integration testing extension.

The integration testing library is written in Python 3.3.
It has only been tested on Linux, but should be portable.
It requires a version of the ENIGMA plugin which supports
changing configurations; if not present, the run tests will
all have the "failed_compilation" end status in the test
reports.

A .gitignore file has also been added to ignore the output
folder, .pyc files, and tests/inttestreport.txt.