Fixed texture binding and drawing function decl.
-Fixed texture binding. This fixed the bug where drawing a circle before
drawing a text would make the circle invisible. That was because texture
was bound to 1 while enigma::bound_texture was 0. This broke
texture_use() and texture_reset() which made draw_circle() (and many
other functions which use texture_reset()) to draw a circle with a bound
texture which made it invisible.
-Fixed some types so it actually compiles. Robert will probably break
them again with his scalar changes. We really need to set up those tests
so bugs like these could be noticed before commiting.