Changes ENIGMA.exe and added gcc.ey
-Made changes to ENIGMA.exe so the exe can be in the same directory as
LGM, as well as work without init scripts. This allows compiling ENIGMA
from source when mingw and git is installed seperatly.
The implemented changes were described here:
http://enigma-dev.org/forums/index.php?topic=1439.0
-Added Compilers\windows\gcc.ey just like for other platforms. It is not
generated and is statically included in the windows patch anyway, so I
see no reason for it to not be in git. If this breaks something, then by
all means remove it.

The .exe code is not very pretty and it should probably be changed, but
it works and I tested it with both installing from ENIGMAPortable.exe
and by using the source.