OpenGL Multi-sampling Implemented the display reset function to OpenGL graphics systems.
Currently it has 3 anomalies.
1) It may not scale properly with views enabled
2) Linux bridge does not have EnableDrawing so I just defaulted display_aa
to 14 or in other words 2,4,and 8 multi-sample levels supported.
3) Same as number two except on Windows I can query in EnableDrawing like
I do with Direct3D except it reports the maximum level and not all the
levels supported, so if you support 2,4,and 8 instead of being 14 like it
should it will report 8.

The bitwise idea comes from Stupido, possibly the only smart thing they've
ever done. At any rate, tested on multiple games and confirmed working,
please merge.