Fixes for Minecraft and Mario
-Fixed more matrix math functions.
-Fixed model draw function which was still using glTranslate.
-Removed all the debug printing.
-Final matrix multiplication is done just before drawing and loading
matrices are also done just before that. On one side this reduces the
useless calls in d3d_ functions. On the other, sometimes a hundred draw
calls are made before a transformation. So there will have to be a check
here to know when to update and when not to update.

In my benchmarks the current GL3 with matrices are 70FPS faster in
Minecraft example (920FPS vs 850FPS), but is slower in Mario examples
because it updates on every draw call. Will add a check and see where it
gets us.