Boosted GL1
-Remade GL1 transformation functions to use the new optimized matrix
functions.

This increases Mario example FPS from 940 to 1190 - thats 20% boost. The
current method in Master (using only GL matrix functions) give 1140FPS,
so the overall boost is much smaller, only about 4%. In examples like
Minecraft, there is no real difference between the old GL1 and this one.
So it's guranteed to be at least as fast, but often times faster (if
using a lot of transformation functions).