GL3 now use shaders.. Finally!
-GL3 now uses shaders for rendering. Shader code for now is in
GL3shader.cpp. The framerate is a lot lower now though, and that is
mostly because of the glGet functions which cause pipeline to stall.
They will be removed, but I have to remake the shader struct for that.

Shader for now is also very simple. I will have to update it to have
lights (missing now), fog, blending and so on.

The whole FFP is gone now. In examples like Minecraft there is only 8
deprecated calls per frame (instead of 8000) and they are to the
glColor4ubv for some reason. I will remove every single on of those
deprecated function calls that lie here and there. But the grunt work is
done.