Fixes ENIGMA (min/max errors and overloads)
Finally tracked down the bug causing min/max not to be recognized and
overloads to be ambiguous (basically also made sin/cos and any other
function to not work).

The bug was in Roberts addition of "using namespace std;" inside
PFthreads.h. He used it just to use deque in one place which is stupid
(using "using namespace" is bad practice in general, I guess just
because of this reason).
As the #include for var4 was after "using namespace std;", then I guess
it was included in that namespace as well. That was probably the cause
for overloads to break.
Now why min/max wasn't recognized by the plugin just because of this
change is a mystery. But I guess namespace again made it go apeshit.