Fixed CPU usage when out of focus
Fixes the bug that CPU usage increase dramatically when the game is
frozen. The game is frozen when out of focus by default. So if I have an
empty game that takes like 1%, when I make it loose focus it goes to 17%
(probably at least 1 of my 8 cores on max). It looses focus also when
minimized, so minimizing a game with "Freeze the game when it looses
focus" option makes it jump to 17%.

This happens because the "freeze" thing happens in main loop and is
checked with:

>if (enigma::gameFroze) continue;

This basically means that it's running an empty loop as fast as
possible, thus using all the resources it can. It can be fixed by doing
adding a delay in the loop. I added 100000us, which means it will only
run the loop 10 times a second. We can increase/decrease the delay if
needed. But now the game uses 0% of CPU when frozen or minimized.