Fixed window_set_pos and implemented minimization
-Fixed window_set_position(). It used wrong window handler. Some users
still use the wrong one. Why are there 2 handlers in that file?
-Added window_set_minimized() and window_get_minimized().

window_set_visible seems broken and in GM hides the window from both the
screen and the task bar (in Win8 you can only see the window in task
manager). In ENIGMA it just show a black window. To make it work the
same as GM, we need to use ShowWindow() with HIDE parameter instead of a
style.