Started on caching uniforms/attributes
First iteration on caching uniforms and attributes.
Now there are std::map's in shader program struct that links names to
locations (so when calling glsl_get_uniform_location we don't have to
call the slow GL function) and another map to link locations to uniform
structs. These structs hold information about the uniform (very useful
for debugging) as well as the content of the uniform. So while the
content is updated we can cache it and get rid of many redundant
glUniform calls.
As we hold information about the type (scalar, vec2, vec3 etc.) as well
as the data type (int, float) then we can print useful debug messages
when a user uses the wrong glsl_uniform function with the given uniform.
The same with attributes, but they are not cached yet.

Todo: Make attributes also cached and add functions for matrix uniforms.