Fixes GL3 lights
-Fixed GL3 lights and at least in Mario example they work the same way
as in GL1. It required some changes in both the shader and the light
system.
-Made light system faster - now it only updates lights that need
updating. I still need to finish the cache system for this to work
effectivly. But even now it greatly increases the performance.
Previously with broken lights GL3 was down to 170FPS in Mario example,
now with working lights it's back to 950FPS.

There could still be problems with point lights. But Mario does seem to
look ok (altough he seems a little brighter in the menu).