Fixed string_width_ext_line and in turn alignment
string_width_ext_line was broken because of the new Unicode changes and
as that function is used when drawing centered text, then that was
broken as well.
It was written 3 years ago and I now see how it might be done better, as
the second "for loop" creates a situation where it goes text over twice.
Maybe later I or someone else can fix that.
Also that "; else;" Josh added 3 years ago was funny. :D To fix compiler
warning Josh didn't want to add braces, so he added a second else.