Fixes GL3 for AMD
The only fix I found was to send color attribute always, even if no
color is specified. This should in turn be slower, so I don't want this
as a permanent fix. Is it possible that color is send and yet
useColor=false? Especially considering it seems that only models were
affected (i.e., the ones loaded by d3d_model_load and drawn with
d3d_model_draw). Maybe the bug is there?