Better GL3 error reporting
-Program class now has "name" member. It can be set manually (in case
you load shaders at runtime) or it's set automatically to the name of
the shader when loaded trough LGM. This means the errors now not only
show ID of the program, but also the name.

Program[0] - Error while linking
Is now:
Program[shr_my_shader = 0] - Error while linking

-Line number shows correctly now. It ignores the prefix part, so if the
error is "error 0:50", then now it is the 50th line in the user shader
and not in the total shader size (which would be prefix + the custom
shader).