Fixes wrong argument to glTexture2D and cleanup
-The internal format parameter for glTexture2D was just 4... nothing
more or less - just the number "4". Dunno who did it or why, but GL spec
requires an enum there. There a lot to choose from (most obvious of
which is GL_RGBA in our case), but none of them actually have the value
4. So the reason the function worked is because when given wrong or
incompatible format, it just chooses one it likes better. It's more of a
"hint".
-Removed deprecated glLight* call.
-Set default ambient to 0,0,0 (c_black) as it's the default one in GM.
Previously we had the default one in GL1 FFP, which is 0.2,0.2,0.2.