Added BasicGUI extension
-This only supports buttons and windows so far, as well as skinning. I
wouldn't want to commit it yet, but this will greatly reduce problems of
me having 4 ENIGMA's installed.
-Examples comming later, but one you can try is this:
https://www.dropbox.com/s/dghg2b6c5wxsbw1/gui_example.egm?dl=0

It will never be extreamlly complicated and functional. It's meant as a
"basic" GUI which holds basic widget components.
NOTE: Using this is not recommened now, as the functions can change a
lot in the future. It's possible that it will be the first thing to use
custom variant types and then many current functions will dissapear.