Two steps forward, one step back
-While implementing sliders I noticed that I actually need two pairs of
functions for sprites, padding, borders and so on. This meant even more
code duplication. So I went the unity way and implemented basis for
styles. Styles define the look of an element. So now all the of the
sprites, padding, fonts, colors and so on are done on styles. Then a
style is assigned to a widget.
-Redid some loops into C++11 style, which actually sped them up about
2-4x.