Worked more on event propagation and groups
-Added groups - they allow things like radio buttons. You create a group
and add several buttons or toggles to it. When one of the buttons is
pressed, then the rest are set to inactive. So like a radio button list.
-Worked more on event propagation, so now it works better when you
activate one element, but then move mouse very rapidly over another.
Right now the depth is broken for some reason, so the events are
reversed.
-Added functions to get if slider, toggle or button is active (pressed).
-Changed many loops to ranged ones just for cleanup.