Function cleanups and implements for undefined ones
-Fixed double define for draw_spline_part with wrong arguments.
-Removed glsl_program_bind_frag_data from header. It was never
implemented and I cannot even find out what it is (it's not a GM
function either).
-Implemented font_get_glyph_left/top/right/bottom. They were fined but
not implemented.
-Added definitions for font_get_glyph_texture_left/top/right/bottom.
They were implemented, but not defined.
-Implemented triangle_area, which was defined, but not implemented.
-Implemented sound_get_pan and sound_get_volume in OpenAL, both of which
were defined.
-Commented out matrix_ functions and d3d_transform_vertex  which were
not implemented.
-Commented up export_include_file,  discard_include_file and
include_file_location as they were not implemented.
-Added empty functions for d3d_set_software_vertex_processing in GL.
Software processing is idiotic anyway, but D3D supports it, and I need
to make a stub until we make platform specific functions easier to
implement.
-Fixed d3d_model_part_draw() definitions - they missed vertex_start
argument.
-Implemented display_get_gui_height and display_get_gui_width.
-Commented out display_get_orientation as it hasn't been implemented (we
don't even support devices with orientation right now).
-Implemented date_get_week and date_inc_week. Apparently I missed it
when I wrote the thing in 2011. 4 years later, it's in (though not ISO).
-Commented out joystick_map_button and joystick_map_axis from
PFjoystick.h, as they are not implemented in Windows, but they are added
in Linux. As they are defined in LINUXjoystick.h, then I guess they
still should work on Linux.
-Implemented mp_grid_clear_cell which was clearly missing.
-Implemented sprite_get/set_bbox_mode. This was also done in the .dll.