Texture atlas.. finally
-ENIGMA is slowly stepping into 21st century - I started on texture
atlases. It actually wasn't that hard to do.
--First the texture atlas will be a runtime thing until LGM (or some
other IDE) will support them. But it's alright with me, as it can be
created on startup.
--Second right now I only added sprite support and they have to be
placed manually. I will work on packing algorithm shortly (I'll probably
the the rect packer made by Josh that is already used in ENIGMA's font
packing).
--Third, as we already do batching, then this give instant rewards. You
just put all of your sprites in an atlas and draw calls reduce to 12. :D
I will post some benchmarks when I'm finished. But in short - worst case
scenario of drawing 1000 images with random of four sprites gave 88fps
previously on my laptop, not it gives 970fps.
--Four, this might break manual draw_texture stuff when the texture is
merged into an atlas. But it's okay. I'll just give functions to get the
position of the texture in atlas. Other than that, this is totally
backwards compatible and shouldn't break anything.

-Built-in shader is now a C++11 raw literal. This makes it easier to
maintain and copy, as we don't need those damn quotes and newlines.
-Fixed bug in the new surface creation, where I actually increment
surface_max count even though the id itself was reused.
-Added graphics_copy_texture() which is required for texture atlases.
Made it for GL3 now, but I will do GL1 later. And then see if I can do
DX9.