Texture atlas automatic packing done
-Implemented automatic packing algorithm based on Josh's rectpacker
already used in ENIGMA. This automatically packs and resizes the texture
atlas if needed. If given a sprite, then all sprite subimages are
packed. I'd say it packs quite efficiently, but not sure how the
algorithm actually works - a lot of pointer magic going on. I think it's
some graph based algorithm.
-Implemented texture packing for fonts too. The idea is that you will be
able to specify all three graphical resources of ENIGMA (sprites,
backgrounds and fonts) and it will be able to pack all of them in one.
Fonts now work fine too, but in some cases some of the letters are cut
off. Next step is to implement backgrounds.
-Texture atlas can have default size (-1,-1) now, which means it will
automatically expand based on how much you want to pack.
-Added graphics_copy_texture_part. This is also needed for texture
atlas, as I need a way to copy only part of the source texture in fonts.
-Fixed memory leak in graphics_copy_texture.
-graphics_copy_texture and the new graphics_copy_texture_part now
correctly crops the image.
-It was actually correct to have "hard coded" 64x64 size in
fontstruct.cpp. That size is not the size of a glyph, but the initial
size of the font texture. If the initial size is taken as the first
glyph size (like 7x9) then resizing will give wrong results. It would
make non-power of two textures. Not a big problem on any modern
hardware, but still not intended in this case.