Added texture atlas to GL1 and BGUI improvements
-GL1 now compiles. I added the necessary texture_ functions and so
texture atlas also works. The performance improvement in GL1 are smaller
than in GL3, because we don't do any batching in GL1. But it can still
improve FPS 2-3x in the worse case scenarios.
-Scrollbar _indicator_size function now limits size in range 0 to 1. It
makes more sense than to return and do nothing.
-Buttons now check for window stencil tests when clicked on. So if the
button is cut (partially or fully) during the stencil pass, it will only
respond with the visible part. This also needs to be done with the rest
of the widgets.