Started on element order
-Started on element ordering, which will allow arbitrary render order
for widget elements in windows.
-Added the generic gui_window_add_element, _remove_element and
_get_element, and element_count functions.
-Added gui_element_get_type function which returns type of the element
with given ID. Useful, as we no longer keep track of each individual
element. For example, this function replaces
gui_window_get_slider_count() and gui_window_get_slider().
-Added new macros, like check_element_exists that doesn't care about the
element type, only that it exists. And get_element_smartv that figures
out the type of the element automatically, which is required for generic
functions.